lib: Move some code from cockpit.js to cockpit/
So that the upcoming SessionManager can use it without depending on
cockpit.js.
